taking a break.  Move to tjb

+++ web/annotate/annrest.py
@@ -16,6 +16,8 @@ import dbconfig
 import annproj
 import h5ann
 
+import operator
+
 from annerror import ANNError
 from pprint import pprint
 
@@ -511,11 +513,13 @@ def getAnnotation ( webargs ):
     if (xmax-xmin)*(ymax-ymin)*(zmax-zmin) >= 1024*1024*16 :
       raise ANNError ("Cutout region is inappropriately large.  Dimension: %s,%s,%s" % (str(xmax-xmin),str(ymax-ymin),str(zmax-zmin)))
 
-    cutoutdata = np.zeros([zmax-zmin+1,ymax-ymin+1,xmax-xmin+1])
+    cutoutdata = np.zeros([zmax-zmin+1,ymax-ymin+1,xmax-xmin+1], dtype=np.uint32)
 
     # rewrite as a list comprehension
     for (a,b,c) in voxarray: 
        cutoutdata[c-zmin,b-ymin,a-xmin] = annoid 
+
+    # try to make more efficient (map with setitem doesn't work).
 
     h5.addCutout ( resolution, [xmin,ymin,zmin], cutoutdata )
